+++ app/components/shadcn-ui/$collapsible.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import * as CollapsiblePrimitive from "@radix-ui/react-collapsible"
 
 function Collapsible({
