+++ app/components/shadcn-ui/$separator.tsx
@@ -1,5 +1,3 @@
-"use client"
-
 import * as SeparatorPrimitive from "@radix-ui/react-separator"
 import type * as React from "react"
